Add tests for group -> project relationships
