+++ src/steps/__tests__/__fixtures__/groupProjectEntities.json
@@ -1,0 +1,27 @@
+[
+  {
+    "id": "gitlab-group:1",
+    "_key": "gitlab-group:1",
+    "_type": "gitlab_group",
+    "_class": ["Group"],
+    "name": "Group 1",
+    "createdOn": 1586982455921
+  },
+  {
+    "id": "gitlab-group:2",
+    "_key": "gitlab-group:2",
+    "_type": "gitlab_group",
+    "_class": ["Group"],
+    "name": "Group 2",
+    "parentGroupId": "gitlab-group:1",
+    "createdOn": 1586982455921
+  },
+  {
+    "id": "gitlab-project:1",
+    "_key": "gitlab-project:1",
+    "_type": "gitlab_project",
+    "_class": ["Project"],
+    "name": "Project 1",
+    "createdOn": 1586982455921
+  }
+]
